refactor(types): separate zoom and color layer contracts
